requirements.txt
numpy==1.26.4
pandas==2.2.1
pycox==0.3.0
shap==0.45.0
scikit-learn==1.3.2
ucimlrepo==0.0.6
tqdm==4.66.2
lifelines==0.28.0
scikit-survival==0.22.2
xgboost==2.0.3
matplotlib==3.8.3
mrmr-selection==0.2.8
Boruta==0.3
skrebate==0.62
torch==2.4.1
seaborn==0.13.2
# git+ssh://git@github.com/O-T-O-Z/FeatBoost-X.git@v0.1.1
scipy==1.13
